RFC: Try harder to remove efivarfs entries we've failed to actually set.
Right now we can get a situation where the firmware has denied a write
via SetVariable() for a variety of reasons.  In such a case, the strace
from userland looks like:

  access("/sys/firmware/efi/efivars/SysPrep0009-8be4df61-93ca-11d2-aa0d-00e098032b8c", F_OK) = -1 ENOENT (No such file or directory)
  open("/sys/firmware/efi/efivars/SysPrep0009-8be4df61-93ca-11d2-aa0d-00e098032b8c", O_WRONLY|O_CREAT, 0644) = 3
  ioctl(3, FS_IOC_GETFLAGS, 0x7ffdee8b8ebc) = 0
  ioctl(3, FS_IOC_SETFLAGS, 0x7ffdee8b8ebc) = 0
  write(3, "\7\0\0\0\1\0\0\0\\\0F\0e\0d\0o\0r\0a\0\0\0\4\1*\0\1\0\0\0\0\10\0\0\0\0\0\0\0@\6\0\0\0\0\0\33&\365,\230{\300H\256TF=\275#\346[\2\2\4\4.\0\\\0E\0F\0I\0\\\0f\0e\0d\0o\0r\0a\0\\\0s\0h\0i\0m\0.\0e\0f\0i\0\0\0\177\377\4\0", 116) = -1 EINVAL (Invalid argument)
  unlink("/sys/firmware/efi/efivars/SysPrep0009-8be4df61-93ca-11d2-aa0d-00e098032b8c") = -1 EINVAL (Invalid argument)

In this case unlink is returning EINVAL because we have tried to delete
the variable through SetVariable, but no variable was ever created.  We
then go on to fail to remove it from the file system as a result.  These
are both the result of our handling of the write error.  Our current
implementation of efivarfs_file_write() has:

        bytes = efivar_entry_set_get_size(var, attributes, &datasize,
                                          data, &set);
        if (!set && bytes) {
                if (bytes == -ENOENT)
                        bytes = -EIO;
                goto out;
        }

        if (bytes == -ENOENT) {
                drop_nlink(inode);
                d_delete(file->f_path.dentry);
                dput(file->f_path.dentry);
        } else {
                inode_lock(inode);
                i_size_write(inode, datasize + sizeof(attributes));
                inode_unlock(inode);
        }

        bytes = count;
 out:

And in most cases we never correctly test for bytes == -ENOENT, because we
have clobbered the error value *and* skipped the test.  It is feasible
this scenario could come from any other error returned from
efivar_entry_set_get_size() as well.

This patch removes that error check, and replaces it with a flag,
"needs_write", which we to true when we create a variable.  If we ever
have a successful write, the flag is set to false.  When the file is
eventually closed, if no successful write has ever happened, we prune it
from our filesystem.