Allow loading kernel above 4G on 64-bit EFI machines.
Previously, on x86_64 hardware, if the kernel was loaded above 4G by the
bootloader and the 64-bit EFI sub entry point was used,

On x86 CPUs, when the kernel is booted via the EFI stub boot loader
entry point, the kernel compares hdr->pref_address to hdr->code32_start.
If those values differ, the kernel attempts to relocate itself to the
address at hdr->pref_address.

Unfortunately, hdr->code32_start is a 32-bit value, and so an address
above 4G cannot be correctly expressed.  As a result, the kernel will
relocate the (likely uninitialized) memory at the low-order 32-bits of
whatever value is in hdr->code32_start, and jumps into that.

That isn't good.

This patch renames code32_start to kernel_start, because I felt stupid
typing ext_code32_start repeatedly, and adds ext_kernel_start to hold
the high order bits.

Signed-off-by: Peter Jones <pjones@redhat.com>