efi: Remove UGA IO as a config table.
The commit at
https://git.kernel.org/pub/scm/linux/kernel/git/tglx/history.git/commit/?id=c596442acc26c3195a69ff63a3f8e6bb12ac4a7d
added all of x86 EFI support in one fell swoop, and when it did so it
added the EFI UGA IO Protocol GUID to the list of config tables.

I'm pretty sure it's never actually been a config table on any system,
and the code in eboot.c that uses UGA_IO_PROTOCOL_GUID doesn't care to
look at efi.uga to find it (because it won't be there anyway...).  It's
time to get rid of it.

Signed-off-by: Peter Jones <pjones@redhat.com>