EFI: try to force Reclaim for the variable store.
One common problem we see in QA with UEFI machines is that the firmware
vendor has not included a Reclaim() call in the runtime SMM backend for
SetVariable(), and so it cannot create new variables if the backings
store is filled with deleted variables and deletion records that have
not been reclaimed.

This is exacerbated by the policy coded into Tiano, which will only
reclaim flash blocks for the OS to use at boot if the amount of storage
space available is *less* than the advertized maximum size for a
variable:

  if (((RemainingCommonRuntimeVariableSpace < mVariableModuleGlobal->MaxVariableSize) ||
       (RemainingCommonRuntimeVariableSpace < mVariableModuleGlobal->MaxAuthVariableSize)) ||
      ...

The result of this is when OS installers remove vestigial Boot####
entries or replace BootOrder, as they must do on any install, the
removal creates a deletion record, and there's no longer space to store
the replacement variable.  This especially happens in QA, because
installing the machine over and over forces the number closer and closer
to the threshold.  Since BootOrder is typically fairly small, that
increases the chances that eventually it will be just the right amount
to trigger this thresholding error.

Ultimately this gets reported as an installer bug, then passed around as
a tools bug and a kernel bug until somebody recognizes it and tells them
to test again.  Upun rebooting to test again, the firmware does the
Reclaim() call and frees up all the deletion records and old variables
from prior installs, and the next attempt at installing works just fine.

This patch attempts to detect if we're getting near the top of the flash
storage, before ExitBootServices() has been called, and attempts to
create and delete a 4096-byte variable at that time, triggering a
call to Reclaim().