efi: use const* paramaters for get/setvar by-ref arguments annotated as IN
The EFI spec does not use const at all in its description of the C
prototypes of EFI services and protocols. However, it does annotate
formal parameters with IN or OUT, where the IN annotation on a by-ref
parameter means that the implementation will not modify the memory
that the argument points at.

The lack of const annotations means that the compiler must assume that
literal strings and GUIDs might be modified in memory, forcing it to
allocate them in a writable section of memory. Given that GUIDs are
the force that holds the EFI universe together, writable GUIDs are
best avoided when possible.

So let's update the set_variable and get_variable function pointer
typedefs to qualify their name and vendor arguments as const, and
update all users of these types as appropriate. This permits the
compiler to put these literals in .rodata, without having to rely
on hacks such as get_efi_var/set_efi_var that just cast away the
type's constness.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>